requirements.txt
aiogram==2.11.2
PyYAML==5.4.1
peewee==3.14.1